+++ src/main/java/org/opentmf/camunda/test/listener/BpmnTaskParseListener.java
@@ -2,12 +2,12 @@ package org.opentmf.camunda.test.listener;
 
 import org.opentmf.camunda.test.helper.ReceiveTaskManager;
 import java.util.Arrays;
-import org.camunda.bpm.engine.delegate.ExecutionListener;
-import org.camunda.bpm.engine.impl.bpmn.parser.AbstractBpmnParseListener;
-import org.camunda.bpm.engine.impl.core.variable.mapping.IoMapping;
-import org.camunda.bpm.engine.impl.pvm.process.ActivityImpl;
-import org.camunda.bpm.engine.impl.pvm.process.ScopeImpl;
-import org.camunda.bpm.engine.impl.util.xml.Element;
+import org.cibseven.bpm.engine.delegate.ExecutionListener;
+import org.cibseven.bpm.engine.impl.bpmn.parser.AbstractBpmnParseListener;
+import org.cibseven.bpm.engine.impl.core.variable.mapping.IoMapping;
+import org.cibseven.bpm.engine.impl.pvm.process.ActivityImpl;
+import org.cibseven.bpm.engine.impl.pvm.process.ScopeImpl;
+import org.cibseven.bpm.engine.impl.util.xml.Element;
 
 /**
  * The `BpmnTaskParseListener` class is a custom implementation of the `AbstractBpmnParseListener`,