+++ src/test/java/org/opentmf/camunda/test/bpm/task/CreateIncidentTask.java
@@ -2,10 +2,10 @@ package org.opentmf.camunda.test.bpm.task;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.camunda.bpm.client.spring.annotation.ExternalTaskSubscription;
-import org.camunda.bpm.client.task.ExternalTask;
-import org.camunda.bpm.client.task.ExternalTaskHandler;
-import org.camunda.bpm.client.task.ExternalTaskService;
+import org.cibseven.bpm.client.spring.annotation.ExternalTaskSubscription;
+import org.cibseven.bpm.client.task.ExternalTask;
+import org.cibseven.bpm.client.task.ExternalTaskHandler;
+import org.cibseven.bpm.client.task.ExternalTaskService;
 import org.springframework.stereotype.Component;
 
 @Slf4j